[WIP] adding sum into integration

+++ torch/csrc/jit/codegen/cuda/arith.cpp
@@ -400,7 +400,7 @@ static TensorView* newForReduction(
 
 TensorView* reductionOp(
     BinaryOpType reduction_op_type,
-    const std::vector<int>& axes,
+    const std::vector<int64_t>& axes,
     Val* init,
     TensorView* tv) {
   TORCH_CHECK(
@@ -414,9 +414,9 @@ TensorView* reductionOp(
   TORCH_CHECK(tv->nDims() > 0, "Tried to reduce a 0-dim tensor");
 
   std::vector<unsigned int> uint_axes;
-  for (int axis : axes) {
+  for (int64_t axis : axes) {
     if (axis < 0)
-      axis += int(tv->nDims());
+      axis += int64_t(tv->nDims());
 
     TORCH_CHECK(
         axis >= 0 && (unsigned int)axis < tv->nDims(),
@@ -436,7 +436,7 @@ TensorView* reductionOp(
   return out;
 }
 
-TORCH_CUDA_API TensorView* sum(TensorView* v1, const std::vector<int>& axes) {
+TORCH_CUDA_API TensorView* sum(TensorView* v1, const std::vector<int64_t>& axes) {
   Val* init;
   switch (v1->getDataType().value()) {
     case (DataType::Float):
